fix(types): mypy 全量通过 + 安全注释修正
- 修复 22 处 mypy 类型错误，245 个源文件 0 错误：
  * _TYPE_MAP 类型标注 (type | tuple[type, ...])
  * httpx/requests Response 统一为 Any
  * openai transcriptions.create call-overload 抑制
  * self_consistency max(key=) 改用 lambda
  * TypedDict → dict 显式转换 (nodes/graph)
  * web/server resp 变量重定义重命名
  * langgraph/frontmatter/jinja2 缺失 stubs 加 import-not-found
  * 内部 compliance/i18n 子模块 import-untyped 抑制
- mcp_server exec(handler_code) 注释修正：
  原注释"限制 exec 内建"与实际不符，改为如实说明
  安全边界（admin-token + feature-flag 双门控）
- 测试：2618 passed, 1 skipped

+++ .traecli/src/deadman/observability/drift.py
@@ -400,9 +400,9 @@ class DriftDetector:
         return words
 
     @staticmethod
-    def _word_freq(words: list[str]) -> dict[str, int]:
+    def _word_freq(words: list[str]) -> dict[str, float]:
         """词频统计"""
-        freq: dict[str, int] = {}
+        freq: dict[str, float] = {}
         for w in words:
             freq[w] = freq.get(w, 0) + 1
         return freq
